requirements.txt
absl-py==0.12.0
aiohttp==3.6.2
altair==4.1.0
appdirs==1.4.4
argon2-cffi==20.1.0
asgiref==3.2.10
astor==0.8.1
astroid==2.4.1
astunparse==1.6.3
async-generator==1.10
async-timeout==3.0.1
attrs==20.2.0
backcall==0.2.0
base58==2.1.0
beautifulsoup4==4.9.1
bleach==3.3.0
blinker==1.4
branca==0.4.2
bs4==0.0.1
cachetools==4.2.1
certifi==2020.4.5.1
cffi==1.14.5
chardet==3.0.4
click==7.1.2
click-plugins==1.1.1
cligj==0.7.1
collection==0.1.6
colorama==0.4.3
cycler==0.10.0
decorator==4.4.2
defusedxml==0.7.1
distlib==0.3.0
Django==3.1
django-bootstrap4==2.3.1
django-cors-headers==3.10.1
django-widget-tweaks==1.4.8
djangorestframework==3.12.2
djangorestframework-simplejwt==5.0.0
entrypoints==0.3
filelock==3.0.12
Flask==1.1.2
Flask-MySQLdb==0.2.0
flatbuffers==1.12
folium==0.12.1
future==0.18.2
gast==0.3.3
geocoder==1.38.1
geographiclib==1.50
geoip2==4.0.2
geopy==2.0.0
gitdb==4.0.7
GitPython==3.1.17
google-auth==1.27.1
google-auth-oauthlib==0.4.3
google-pasta==0.2.0
grpcio==1.32.0
h5py==2.10.0
html5lib==1.0.1
idna==2.9
image==1.5.33
ipykernel==5.5.5
ipython==7.24.0
ipython-genutils==0.2.0
ipywidgets==7.6.3
isort==4.3.21
itsdangerous==1.1.0
jedi==0.18.0
Jinja2==2.11.2
joblib==1.0.1
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.1.12
jupyter-console==6.4.0
jupyter-core==4.7.1
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.0
keras==2.7.0
Keras-Preprocessing==1.1.2
kiwisolver==1.2.0
lazy-object-proxy==1.4.3
libclang==12.0.0
lxml==4.6.3
Markdown==3.3.4
MarkupSafe==1.1.1
matplotlib==3.3.0
matplotlib-inline==0.1.2
maxminddb==2.0.2
mccabe==0.6.1
mediapipe==0.8.4.2
mistune==0.8.4
more-itertools==8.8.0
multidict==4.7.6
multitasking==0.0.9
munch==2.5.0
mysql==0.0.2
mysql-connector==2.2.9
mysqlclient==2.0.1
nbclient==0.5.3
nbconvert==6.0.7
nbformat==5.1.3
nest-asyncio==1.5.1
notebook==6.4.0
numpy==1.19.5
oauthlib==3.1.0
opencv-contrib-python==4.5.2.52
opencv-python==4.3.0.36
opt-einsum==3.3.0
packaging==20.9
pandas==1.2.3
pandocfilters==1.4.3
parso==0.8.2
pickleshare==0.7.5
Pillow==7.2.0
plotly==4.14.3
prettytable==2.0.0
prometheus-client==0.10.1
prompt-toolkit==3.0.18
protobuf==3.15.6
pyarrow==4.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pydeck==0.6.2
pygame==2.0.1
Pygments==2.9.0
PyJWT==2.3.0
pylint==2.5.2
PyMySQL==1.0.2
pyparsing==2.4.7
pyproj==3.0.0.post1
pyrsistent==0.17.3
python-dateutil==2.8.1
python-decouple==3.4
pytz==2020.1
pywin32==300
pywinpty==1.1.1
PyYAML==5.4.1
pyzmq==22.1.0
qtconsole==5.1.0
QtPy==1.9.0
random2==1.0.1
ratelim==0.1.6
requests==2.24.0
requests-oauthlib==1.3.0
retrying==1.3.3
rsa==4.7.2
scikit-learn==0.24.1
scipy==1.6.1
selenium==3.141.0
Send2Trash==1.5.0
Shapely==1.7.1
six==1.15.0
smmap==4.0.0
soupsieve==2.0.1
sqlparse==0.3.1
streamlit==0.82.0
tensorboard==2.7.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorflow==2.4.1
tensorflow-estimator==2.7.0
tensorflow-gpu==2.7.0
tensorflow-io-gcs-filesystem==0.22.0
termcolor==1.1.0
terminado==0.10.0
testpath==0.5.0
threadpoolctl==2.1.0
tk==0.1.0
toml==0.10.1
toolz==0.11.1
tornado==6.1
traitlets==5.0.5
twilio==6.44.2
typing-extensions==3.7.4.3
tzlocal==2.1
urllib3==1.25.9
validators==0.18.2
virtualenv==20.0.21
watchdog==2.1.2
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==1.0.1
widgetsnbextension==3.5.1
wrapt==1.12.1
yarl==1.5.1
yfinance==0.1.59
